fix(device-quota): address phase 5 review findings
